Edit/Delete Access issue fix
Earlier any non-owner of the image can edit/delete images this issue has been resolved

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
@@ -92,13 +93,23 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session, RedirectAttributes redirect) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        User currentUser = (User) session.getAttribute("loggeduser");
+        
+        if(imageService.isImageOwner(image, currentUser)) {
+        	String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        else {
+        	String imageTitle = imageService.getImage(imageId).getTitle();
+        	String error = "Only the owner of the image can edit the image";
+        	redirect.addFlashAttribute("editError", error);
+        	return "redirect:/images/" + imageId + '/' + imageTitle;
+        }
+        
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -132,7 +143,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" + imageId + '/' + updatedImage.getTitle();
     }
 
 
@@ -140,9 +151,21 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session,RedirectAttributes redirect) {
+    	 Image image = imageService.getImage(imageId);
+         User currentUser = (User) session.getAttribute("loggeduser");
+         
+         if(imageService.isImageOwner(image, currentUser)) {
+        	  imageService.deleteImage(imageId);
+        	  return "redirect:/images";
+         }
+         else {
+        	  String imageTitle = imageService.getImage(imageId).getTitle();
+              String error = "Only the owner of the image can delete the image";
+              redirect.addFlashAttribute("deleteError", error);
+              return "redirect:/images/" + imageId + '/' + imageTitle;
+         }
+       
     }
